half working cookbook

+++ recipes/default.rb
@@ -3,22 +3,29 @@
 # Recipe:: default
 #
 # Copyright (C) 2012 Lucas Jandrew
-# 
+#
 # Licensed under the Apache License, Version 2.0 (the "License");
 # you may not use this file except in compliance with the License.
 # You may obtain a copy of the License at
-# 
+#
 #    http://www.apache.org/licenses/LICENSE-2.0
-# 
+#
 # Unless required by applicable law or agreed to in writing, software
 # distributed under the License is distributed on an "AS IS" BASIS,
 # WITHOUT WARRANTIES OR CONDITIONS OF ANY KIND, either express or implied.
 # See the License for the specific language governing permissions and
 # limitations under the License.
 #
-package 'libev-dev'
-package 'libssl-dev'
-package 'git'
+include_recipe "build-essential"
+
+platform_options = node["stud"]["platform"]
+
+platform_options["stud_packages"].each do |pkg|
+  package pkg do
+    action :install
+      options platform_options["package_overrides"]
+  end
+end
 
 directory '/opt/stud/build' do
   mode '0755'
@@ -34,13 +41,7 @@ directory '/etc/stud/certs' do
   recursive true
 end
 
-directory '/etc/stud/configs-available' do
-  mode '0755'
-  owner 'root'
-  group 'root'
-end
-
-directory '/etc/stud/configs-enabled' do
+directory '/etc/stud/conf.d' do
   mode '0755'
   owner 'root'
   group 'root'
@@ -55,10 +56,18 @@ group 'stud' do
   system true
 end
 
+###################################################
+### NOTE:(mancdaz) until we get packaging sorted ##
 git '/opt/stud/build' do
   repository 'https://github.com/bumptech/stud.git'
   reference 'master'
   action :sync
+end
+
+if node['platform'] == 'redhat'
+  link '/usr/include/ev.h' do
+    to '/usr/include/libev/ev.h'
+  end
 end
 
 bash "build stud" do
@@ -68,16 +77,30 @@ bash "build stud" do
     make install
     mv stud /opt/stud/
   EOH
+  not_if { ::File.exists?("/usr/local/bin/stud") }
 end
+###################################################
 
-template '/etc/init/stud.conf' do
+# upstart config file works on ubuntu and rhel
+cookbook_file '/etc/init/stud.conf' do
   mode '0644'
-  source 'init/stud.conf.erb'
+  source 'stud-upstart.conf'
 end
 
 service 'stud' do
   case node[:platform]
-  when 'ubuntu'
+  when 'ubuntu', 'redhat'
     provider Chef::Provider::Service::Upstart
   end
+  action :enable
+end
+
+
+stud_instance "mything" do
+    pemfile "/tmp/apemfile.pem"
+    source_hostname = '127.0.0.1'
+    source_port = '4000'
+    destination_hostname = '127.0.0.1'
+    destination_port = '22'
+    certificate_domain = 'anameofadomain'
 end
